TST: Reorganize tests of Ad machinery.
Gather all tests relating to the forward mode in one module.
