+++ tests/unit/ad/test_grid_operators.py
@@ -7,6 +7,8 @@
     BoundaryCondition
     ParameterVector
     ParameterMatrix
+
+Tests of the basic Discretization class are also placed here.
 """
 from __future__ import annotations
 
@@ -15,7 +17,6 @@ import pytest
 import scipy.sparse as sps
 
 import porepy as pp
-from tests.unit.test_ad import _compare_matrices, _list_ind_of_grid
 
 
 def set_parameters(
@@ -475,3 +476,172 @@ def test_divergence(mdg: pp.MixedDimensionalGrid, dim: int):
     op = pp.ad.Divergence(subdomains)
     val = op.parse(mdg)
     _compare_matrices(val, sps.block_diag(divergences))
+
+
+
+
+@pytest.mark.parametrize(
+    "grids",
+    [
+        [pp.CartGrid(np.array([4, 1]))],
+        [pp.CartGrid(np.array([4, 1])), pp.CartGrid(np.array([2, 2]))],
+    ],
+)
+@pytest.mark.parametrize(
+    "variables",
+    [["foo"], ["foo", "bar"]],
+)
+def test_variable_combinations(grids, variables):
+    # Test combinations of variables, and merged variables, on different grids.
+    # Main check is if Jacobian matrices are of the right size.
+
+    # Make MixedDimensionalGrid, populate with necessary information
+    mdg = pp.MixedDimensionalGrid()
+    mdg.add_subdomains(grids)
+    for sd, data in mdg.subdomains(return_data=True):
+        data[pp.STATE] = {}
+        data[pp.PRIMARY_VARIABLES] = {}
+        for var in variables:
+            data[pp.PRIMARY_VARIABLES].update({var: {"cells": 1}})
+            data[pp.STATE][var] = np.random.rand(sd.num_cells)
+
+    # Ad boilerplate
+    dof_manager = pp.DofManager(mdg)
+    eq_manager = pp.ad.EquationManager(mdg, dof_manager)
+
+    # Standard Ad variables
+    ad_vars = [eq_manager.variable(g, var) for g in grids for var in variables]
+    # Merge variables over all grids
+    merged_vars = []
+    for var in variables:
+        gv = [(g, var) for g in grids]
+        merged_vars.append(eq_manager.merge_variables(gv))
+
+    # First check of standard variables. If this fails, something is really wrong
+    for sd in grids:
+        data = mdg.subdomain_data(sd)
+        for var in ad_vars:
+            if sd == var._g:
+                expr = var.evaluate(dof_manager)
+                # Check that the size of the variable is correct
+                assert np.allclose(expr.val, data[pp.STATE][var._name])
+                # Check that the Jacobian matrix has the right number of columns
+                assert expr.jac.shape[1] == dof_manager.num_dofs()
+
+    # Next, check that merged variables are handled correctly.
+    for var in merged_vars:
+        expr = var.evaluate(dof_manager)
+        vals = []
+        for sub_var in var.sub_vars:
+            vals.append(mdg.subdomain_data(sub_var._g)[pp.STATE][sub_var._name])
+
+        assert np.allclose(expr.val, np.hstack([v for v in vals]))
+        assert expr.jac.shape[1] == dof_manager.num_dofs()
+
+    # Finally, check that the size of the Jacobian matrix is correct when combining
+    # variables (this will cover both variables and merged variable with the same name,
+    # and with different name).
+    for sd in grids:
+        for var in ad_vars:
+            nc = var.size()
+            cols = np.arange(nc)
+            data = np.ones(nc)
+            for mv in merged_vars:
+                nr = mv.size()
+
+                # The variable must be projected to the full set of grid for addition
+                # to be meaningful. This requires a bit of work.
+                sv_size = np.array([sv.size() for sv in mv.sub_vars])
+                mv_grids = [sv._g for sv in mv.sub_vars]
+                ind = mv_grids.index(var._g)
+                offset = np.hstack((0, np.cumsum(sv_size)))[ind]
+                rows = offset + np.arange(nc)
+                P = pp.ad.Matrix(sps.coo_matrix((data, (rows, cols)), shape=(nr, nc)))
+
+                eq = eq = mv + P * var
+                expr = eq.evaluate(dof_manager)
+                # Jacobian matrix size is set according to the dof manager,
+                assert expr.jac.shape[1] == dof_manager.num_dofs()
+
+
+def test_ad_discretization_class():
+    # Test of the mother class of all discretizations (pp.ad.Discretization)
+
+    fracs = [np.array([[0, 2], [1, 1]]), np.array([[1, 1], [0, 2]])]
+    mdg = pp.meshing.cart_grid(fracs, np.array([2, 2]))
+
+    subdomains = [g for g in mdg.subdomains()]
+    sub_list = subdomains[:2]
+
+    # Make two Mock discretizations, with different keywords
+    key = "foo"
+    sub_key = "bar"
+    discr = _MockDiscretization(key)
+    sub_discr = _MockDiscretization(sub_key)
+
+    # Ad wrappers
+    # This is mimics the old init of Discretization, before it was decided to
+    # make that class semi-ABC. Still checks the wrap method
+    discr_ad = pp.ad.Discretization()
+    discr_ad.subdomains = subdomains
+    discr_ad._discretization = discr
+    pp.ad._ad_utils.wrap_discretization(discr_ad, discr, subdomains)
+    sub_discr_ad = pp.ad.Discretization()
+    sub_discr_ad.subdomains = sub_list
+    sub_discr_ad._discretization = sub_discr
+    pp.ad._ad_utils.wrap_discretization(sub_discr_ad, sub_discr, sub_list)
+
+    # values
+    known_val = np.random.rand(len(subdomains))
+    known_sub_val = np.random.rand(len(sub_list))
+
+    # Assign a value to the discretization matrix, with the right key
+    for vi, sd in enumerate(subdomains):
+        data = mdg.subdomain_data(sd)
+        data[pp.DISCRETIZATION_MATRICES] = {key: {"foobar": known_val[vi]}}
+
+    # Same with submatrix
+    for vi, sd in enumerate(sub_list):
+        data = mdg.subdomain_data(sd)
+        data[pp.DISCRETIZATION_MATRICES].update(
+            {sub_key: {"foobar": known_sub_val[vi]}}
+        )
+
+    # Compare values under parsing. Note we need to pick out the diagonal, due to the
+    # way parsing make block matrices.
+    assert np.allclose(known_val, discr_ad.foobar.parse(mdg).diagonal())
+    assert np.allclose(known_sub_val, sub_discr_ad.foobar.parse(mdg).diagonal())
+
+
+## Below are helpers for tests of the Ad wrappers.
+
+
+def _compare_matrices(m1, m2):
+    if isinstance(m1, pp.ad.Matrix):
+        m1 = m1._mat
+    if isinstance(m2, pp.ad.Matrix):
+        m2 = m2._mat
+    if m1.shape != m2.shape:
+        return False
+    d = m1 - m2
+    if d.data.size > 0:
+        if np.max(np.abs(d.data)) > 1e-10:
+            return False
+    return True
+
+
+def _list_ind_of_grid(subdomains, g):
+    for i, gl in enumerate(subdomains):
+        if g == gl:
+            return i
+
+    raise ValueError("grid is not in list")
+
+
+class _MockDiscretization:
+    def __init__(self, key):
+        self.foobar_matrix_key = "foobar"
+        self.not_matrix_keys = "failed"
+
+        self.keyword = key
+
